pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "yamlguard"
version = "0.1.0"
description = "Local-first YAML Guard (validation + policy rules + API + CLI)"
authors = [{name = "You"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
"fastapi>=0.115",
"uvicorn>=0.30",
"ruamel.yaml>=0.18",
"jsonschema>=4.23",
"jsonpath-ng>=1.6",
"pydantic>=2.7"
]


[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pre-commit>=3.7"
]


[project.scripts]
yamlguard = "yamlguard.cli.main:main"


[tool.pytest.ini_options]
addopts = "-q"